Restructures packages by subjects
