+++ src/main/java/org/jboss/arquillian/extension/jacoco/client/JaCoCoArchiveAppender.java
@@ -21,6 +21,7 @@ import org.jboss.arquillian.container.test.spi.client.deployment.AuxiliaryArchiv
 import org.jboss.arquillian.core.api.Instance;
 import org.jboss.arquillian.core.api.annotation.Inject;
 import org.jboss.arquillian.extension.jacoco.CoverageDataCommand;
+import org.jboss.arquillian.extension.jacoco.client.configuration.JaCoCoConfiguration;
 import org.jboss.arquillian.extension.jacoco.container.JacocoRemoteExtension;
 import org.jboss.arquillian.extension.jacoco.container.StartCoverageData;
 import org.jboss.shrinkwrap.api.Archive;
