+++ src/main/java/org/jboss/arquillian/extension/jacoco/client/JaCoCoExtension.java
@@ -20,6 +20,8 @@ package org.jboss.arquillian.extension.jacoco.client;
 import org.jboss.arquillian.container.test.spi.client.deployment.ApplicationArchiveProcessor;
 import org.jboss.arquillian.container.test.spi.client.deployment.AuxiliaryArchiveAppender;
 import org.jboss.arquillian.core.spi.LoadableExtension;
+import org.jboss.arquillian.extension.jacoco.client.configuration.JaCoCoConfiguration;
+import org.jboss.arquillian.extension.jacoco.client.configuration.JaCoCoConfigurator;
 
 /**
  * JaCoCoExtension
