+++ src/main/java/org/jboss/arquillian/extension/jacoco/client/instrumentation/SignatureRemover.java
@@ -3,7 +3,7 @@
  * All rights reserved.
  *
  */
-package org.jboss.arquillian.extension.jacoco.client;
+package org.jboss.arquillian.extension.jacoco.client.instrumentation;
 
 import java.util.Map;
 import java.util.Map.Entry;
