+++ src/test/java/org/jboss/arquillian/extension/jacoco/test/unit/InstrumentTestCase.java
@@ -30,8 +30,8 @@ import org.jacoco.core.data.ISessionInfoVisitor;
 import org.jacoco.core.data.SessionInfo;
 import org.jacoco.core.instr.Instrumenter;
 import org.jacoco.core.runtime.RuntimeData;
-import org.jboss.arquillian.extension.jacoco.client.InstrumentedAsset;
-import org.jboss.arquillian.extension.jacoco.client.SignatureRemover;
+import org.jboss.arquillian.extension.jacoco.client.instrumentation.InstrumentedAsset;
+import org.jboss.arquillian.extension.jacoco.client.instrumentation.SignatureRemover;
 import org.jboss.arquillian.extension.jacoco.container.ArquillianRuntime;
 import org.jboss.arquillian.extension.jacoco.test.included.CoverageBean;
 import org.jboss.shrinkwrap.api.ArchivePath;
